services/workers/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatif-workers"
version = "0.1.0"
description = "WhatIf async asset workers"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "google-genai>=1.65.0",
  "google-cloud-storage>=2.18.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.6.0"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
